Implement memos and channel commands.
This implements the ability for users to leave memos to other users, to
be delivered via a private message when the recipient next joins the
channel. Afterwards, the memos are cleared from the database.

The parser was changed a bit to accomodate commands with arguments.
Also, since users usually expect to be able to leave memos from the
channel, the bot can now be given commands by highlighting it in the
channel (messages are considered to highlight if they mention the bot's
nickname as the first word of the message, optionally followed by some
suffix, currently ":" or ","), followed by the command. The bot sends
its response to wherever it received the command from (i.e. either the
channel or via a private message to the user), except in the case of
the known_users command where the output is inappropriately long.