Remove known_users command
This removes the known_users command, which is problematic because it
allows everyone to see all users in the database and because it makes
the bot crash when it is used, due to flooding. This command also isn't
needed anymore, because you can just manually inspect the database if
you ever want to debug this data.